requirements.txt
streamlit
pandas
plotly
spacy
google-genai
python-dotenv
pdfplumber
tenacity